Consolodate hlr/nod central mail status logic into concern
